Add the JITLog Message definion parser
It takes an array of message definion tables and for each message it parses all
the field strings in to field definition tables that the writer and reader
generators use.
The tables are built-up in 2 passes, first it figures out what field any
bit\boolean fields get stored in to for now its just the spare bits of
the message header and also try to pack fields smaller than 32 bits in the
header if theres space left.
In a second pass of the field list it calculates offset of fields, the size of
messages and validates field names used for length of array fields exist.