Record when a full trace flushes happens to JITLog
The message includes the reason, current maxtrace/maxmcode limits and timestamp.
Added a enumdef for the flushreason the current reason are:
user_requested, // jit.flush was called
maxmcode,        // Max machine code limit was hit.
maxtrace,          // Max trace count limit was hit.
profile_toggle,    // built-in sample based profiler is toggled on or off.
set_builtinmt,     // The meta table of a built-in type was changed.
set_immutableuv // An immutable upvalue was changed breaking the invariant
that jit'ed code depends on, not currently implemented.

Record the current jitparam values and there defaults when starting the JITLog