Simplify copying many values from an internal structs to a message with a new msg definition field(struct_args)
The list of fields to copy is specified as combined array and hash table.
The target field in the message is implicitly assumed to be the same name as
the struct field name unless specified using a hash table entry.
The address of the struct can also optionally be saved in to a specified
message field using a store_address entry.