Add per function hot counting
Function hot counter have been moved from the shared hashtable to inside the
function prototype object that is only shared between the closures of a
function. The function hot counter field was added to the end of the GCproto
struct so that its right next the function header bytecode in memory which
decrements the counter.

Since the function hot counters are now spread out around the function proto
objects it will increase the number of dirty cache lines that need to
be eventually be flushed back to memory vs only two cache lines of the shared
hotcount hashtable, but this stops happening once the function is JIT'ed.

Updated the interpreter assembly for the Lua function header bytecodes to use
the new hot counter location.
Added a separate JIT param for function hot trigger count currently set to the
same effective value as it was with the shared hotcount table system of 112 calls.