ARM: Fix up condition codes for conditional arithmetic insn
When an arithmetic instruction such as add or sub are combined with a
subsequent compare with zero, its following conditional branch code
needs fixing up.  This is necessary because one could generate an add
with a subtract of the negative but such a substitution, while correct
on its own, will change the effect on condition flags since while
addition of two positive numbers may signal an overflow, addition of a
positive and a negative number may not.  So if earlier the condition
code was GE, it needs to be fixed up to PL to remain correct.

We did that for bit operations but not for arithmetic, so do that now.

Fix #38.

Signed-off-by: Yichun Zhang (agentzh) <yichun@openresty.com>