Add support for saving a GC snapshot to a file
Added gcsnapshot_validate which will check that objects in a gc snapshot
look sane.
Changed creategcdump/gcsnapshot_create API to actually allow
access to snapshot. The memory is still leaked though.

Added gcsnapshot_savetofile that allow saving a GC snapshot to file/already open
file steam in a RIFF chunk format in 2 chunks with the first being GCOB which is
the array of object pointers with there type and size and followed by a OMEM
chunk that contains the raw memory of the objects from the first chunk.